+++ CHANGES.md
@@ -1,5 +1,9 @@
 ### 2.2.2.dev0 - In development
 
+- Add a `set_start_method` function in `loky.backend.context`. Note
+  that now, `loky` does not respect the start method set using
+  `multiprocessing.set_start_method` anymore. It is thus mandatory
+  to use the `loky` function to have the correct behavior.
 
 
 ### 2.2.1 - 2018-08-27 - Bug fix release